src/requirements.txt
whatsapp-api-client-python
google-generativeai